+++ server/src/Events/OnCompletion.ts
@@ -2,6 +2,7 @@ import { TextDocument, CompletionItem, Position, CompletionItemKind } from 'vsco
 import { GlobalAnalyzer, parserFunctions } from '../server';
 import { TextParser } from "./../TextParser";
 import { Script } from '../Script';
+import { CursorPositionType } from '../CursorPositionInformation';
 
 let completionCached :CompletionItem[] = [];
 let posCached :Position|null = null;
@@ -26,9 +27,9 @@ export function OnCompletion(docs :Map<string, TextDocument>, curDoc :TextDocume
 			return parserFunctions.getTableFunctions(pos);
 		}
 		
-		scripttext = GlobalAnalyzer.getCompleteCurrentScript(pos, curDoc, docs, false, true);
+		scripttext = GlobalAnalyzer.getCompleteCurrentScript(pos, curDoc, docs, true, true);
 		if(scripttext) {
-			let varPattern = new RegExp("\\b(int|CString|CTable|double|CMoney|CDateTime|float|BOOL)\\s+" + word.m_context + "\\s*(\\=|\\;)", "g");
+			let varPattern = new RegExp("\\b(int|CString|CTable|double|CMoney|CDateTime|float|BOOL)\\s*(\\&|)\\s*" + word.m_context + "\\s*(\\=|\\;|\\,|\\))", "g");
 			let m = varPattern.exec(scripttext.m_scripttext);
 			
 			if(m) {
@@ -44,8 +45,27 @@ export function OnCompletion(docs :Map<string, TextDocument>, curDoc :TextDocume
 				} 
 			}
 		}
+	} else if(word.m_word == "Call:" || word.m_type == CursorPositionType.USERDEFINED_FUNCTION) {
+		let patternFunction = /\bFUNCTION:\s+(void|double|CString|int|BOOL|CTable|CMoney|CDateTime)\s+(.*)\(.*\).*$/gm;
+		scripttext = GlobalAnalyzer.getCompleteCurrentScript(pos, curDoc, docs, true, true);
+		if(scripttext) {
+			let completionFunction :CompletionItem[] = [];
+			let alreadyAdded :string[] = [];
+			let m :RegExpExecArray|null = null;
+			while(m = patternFunction.exec(scripttext.m_scripttext)) {
+				if(!alreadyAdded.includes(m[2])) {
+					alreadyAdded.push(m[2]);
+					completionFunction.push({
+						label: <string>m[2],
+						commitCharacters: ["("],
+						kind: CompletionItemKind.Function
+					});
+				}
+			}
+			return completionFunction;
+		}
 	} else {
-		scripttext = GlobalAnalyzer.getCompleteCurrentScript(pos, curDoc, docs, false, true);
+		scripttext = GlobalAnalyzer.getCompleteCurrentScript(pos, curDoc, docs, true, true);
 		if(scripttext) {
 			if(posCached) {
 				if(posCached.line == pos.line) {
@@ -58,7 +78,7 @@ export function OnCompletion(docs :Map<string, TextDocument>, curDoc :TextDocume
 			completionCached = [];
 			posCached = pos;
 		
-			let varPatternNew = /\b(int|CString|CTable|double|CMoney|CDateTime|float|BOOL)\s+[a-zA-Z0-9_]+\s*(\=|\;)/g;
+			let varPatternNew = /\b(int|CString|CTable|double|CMoney|CDateTime|float|BOOL)\s*(\&|)\s*[a-zA-Z0-9_]+\s*(\=|\;|\,|\))/g;
 			let text = scripttext.m_scripttext;
 			let m :RegExpExecArray|null = null;
 			while(m = varPatternNew.exec(text)) {
@@ -95,6 +115,10 @@ export function OnCompletion(docs :Map<string, TextDocument>, curDoc :TextDocume
 		}
 	}
 
-	completionCached.push(...parserFunctions.getConstantVariables());
+	parserFunctions.getConstantVariables().forEach(element => {
+		if(!completionCached.includes(element)) {
+			completionCached.push(element);
+		}
+	});
 	return completionCached;
 }